Extraction of SLS Patterns from Protein Structures, Bioinformatics
Studying the SLS from a set of protein structures. An SLS is a segment
of protein structure containing a stand of a beta-sheet followed by a
loop/turn and then followed by another stand. The overall task of this
project was to classify a set of input structures into different groups
of SLS that differ by the number of amino acids in the loop/turn.